fix: room permissions
