+++ packages/app/app/components/Message/Model/Room/Settings/Type/Permissions/Index.vue
@@ -2,7 +2,7 @@
 import type { Room } from "@esposter/db-schema";
 
 import { useRoleStore } from "@/store/message/room/role";
-import { RoomPermission } from "@esposter/shared";
+import { RoomPermission } from "@esposter/db-schema";
 
 interface PermissionsSettingsProps {
   roomId: Room["id"];
