+++ packages/app/server/trpc/routers/room/index.ts
@@ -38,6 +38,7 @@ import {
   invites,
   roomRoles,
   rooms,
+  RoomPermission,
   RoomType,
   selectInviteSchema,
   selectRoomSchema,
@@ -46,14 +47,7 @@ import {
   usersToRooms,
   UserToRoomRelations,
 } from "@esposter/db-schema";
-import {
-  InvalidOperationError,
-  ItemMetadataPropertyNames,
-  NotFoundError,
-  Operation,
-  RoomPermission,
-  takeOne,
-} from "@esposter/shared";
+import { InvalidOperationError, ItemMetadataPropertyNames, NotFoundError, Operation, takeOne } from "@esposter/shared";
 import { TRPCError } from "@trpc/server";
 import { and, count, desc, eq, getTableColumns, ilike, inArray, ne, sql } from "drizzle-orm";
 import { alias } from "drizzle-orm/pg-core";
