+++ packages/db-schema/src/schema/roomRoles.ts
@@ -6,6 +6,25 @@ import { bigint, boolean, check, index, integer, text, uniqueIndex, uuid } from 
 import { createSelectSchema } from "drizzle-zod";
 import { z } from "zod";
 
+/* eslint-disable perfectionist/sort-objects */
+export const RoomPermission = {
+  ReadMessages: 1n,
+  SendMessages: 2n,
+  ManageMessages: 4n,
+  MentionEveryone: 8n,
+  ManageRoom: 16n,
+  ManageRoles: 32n,
+  ManageInvites: 64n,
+  KickMembers: 128n,
+  BanMembers: 256n,
+  MuteMembers: 512n,
+  MoveMembers: 1024n,
+  Administrator: 2048n,
+} as const;
+/* eslint-enable perfectionist/sort-objects */
+
+export type RoomPermission = (typeof RoomPermission)[keyof typeof RoomPermission];
+
 export const ROOM_ROLE_NAME_MAX_LENGTH = 100;
 
 export const roomRoles = pgTable(
